6.22.

+++ ch17_JDBC_4/src/com/yedam/app/products/ProductInfoManagement.java
@@ -1,24 +1,34 @@
 package com.yedam.app.products;
 
+import java.util.List;
+
 import com.yedam.app.common.Management;
 
 public class ProductInfoManagement extends Management {
 
 	// 제품에 대한 정보를 등록, 수정, 삭제
 	public ProductInfoManagement() {
+		boolean role = selectRole();
+
 		while (true) {
-			menuPrint();
+			menuPrint(role);
 			int menu = menuSelect();
 
-			if (menu == 1) {
+			if (menu == 1 && role) {
 				// 제품정보 등록
 				insertProductInfo();
-			} else if (menu == 2) {
+			} else if (menu == 2 && role) {
 				// 제품정보 수정 - 가격
 				updateProductPrice();
-			} else if (menu == 3) {
+			} else if (menu == 3 && role) {
 				// 제품정보 삭제
 				deleteProductInfo();
+			} else if (menu == 4) {
+				// 제품정보 조회
+				selectProductInfo();
+			} else if (menu == 5) {
+				// 전체제품조회
+				selectAllProduct();
 			} else if (menu == 9) {
 				// 뒤로가기
 				back();
@@ -29,6 +39,25 @@ public class ProductInfoManagement extends Management {
 
 		}
 
+	}
+
+	private void selectAllProduct() {
+		List<Product> list = pDAO.selectAll();
+		for (Product p : list) {
+			System.out.println(p);
+		}
+	}
+
+	private void selectProductInfo() {
+		// 제품이름 입력
+		String productName = inputName();
+		// 제품정보 검색
+		Product product = pDAO.selectOneName(productName);
+		if (product == null) {
+			System.out.println("등록된 정보가 없습니다.");
+			return;
+		}
+		System.out.println(product);
 	}
 
 	private void updateProductPrice() {
@@ -50,14 +79,14 @@ public class ProductInfoManagement extends Management {
 	}
 
 	private Product inputUpdateInfo(Product product) {
-		System.out.println("기존 정보");
+		System.out.println("<기존 정보>");
 		System.out.println(product.getProductName() + ", " + product.getProductPrice() + "원");
-		System.out.print("이름(수정하지 않으려면 -1입력)>");
+		System.out.print("이름(수정하지 않으려면 -1입력)> ");
 		String name = sc.nextLine();
 		if (!name.equals("-1")) {
 			product.setProductName(name);
 		}
-		System.out.print("가격(수정하지 않으려면 -1입력)>");
+		System.out.print("가격(수정하지 않으려면 -1입력)> ");
 		int price = Integer.parseInt(sc.nextLine());
 		if (price != -1) {
 			product.setProductPrice(price);
@@ -86,7 +115,7 @@ public class ProductInfoManagement extends Management {
 	}
 
 	private String inputName() {
-		System.out.print("제품이름>");
+		System.out.print("제품이름> ");
 		return sc.nextLine();
 	}
 
@@ -99,21 +128,29 @@ public class ProductInfoManagement extends Management {
 
 	private Product inputAll() {
 		Product product = new Product();
-		System.out.print("제품이름>");
+		System.out.print("제품이름> ");
 		product.setProductName(sc.nextLine());
-		System.out.print("제품가격>");
+		System.out.print("제품가격> ");
 		product.setProductPrice(Integer.parseInt(sc.nextLine()));
 		return product;
 	}
 
-	private void back() {
-		System.out.println("이전화면으로 돌아갑니다.");
+	protected void menuPrint() {
+		System.out.println("--------------------------------------------------------------------------");
+		System.out.println(" 1.제품정보등록  2.제품가격수정  3.제품정보삭제  4.제품정보조회  5.전체조회  9.뒤로가기");
+		System.out.println("--------------------------------------------------------------------------");
 	}
 
 	@Override
-	protected void menuPrint() {
-		System.out.println("--------------------------------------------------");
-		System.out.println(" 1.제품정보등록  2.제품가격수정  3.제품정보삭제  9.뒤로가기");
-		System.out.println("--------------------------------------------------");
+	protected void menuPrint(boolean role) {
+
+		String menu = "";
+		if (role) {
+			menu += " 1.제품정보등록  2.제품가격수정  3.제품정보삭제 ";
+		}
+		menu += " 4.제품정보조회  5.전체조회  9.뒤로가기";
+		System.out.println("--------------------------------------------------------------------------");
+		System.out.println(menu);
+		System.out.println("--------------------------------------------------------------------------");
 	}
 }
